+++ src/main/java/com/yhm/universityhelper/dao/UserRoleMapper.java
@@ -21,5 +21,5 @@ import java.util.List;
 @Mapper
 public interface UserRoleMapper extends BaseMapper<UserRole> {
     @Select("select * from uh_user_role where userId = #{userId}")
-    List<UserRole> listByUserId(Integer userId);
+    List<UserRole> listByUserId(Long userId);
 }
